frontend: Home/ClusterContextMenu: Extract ContextMenu from Home
